+++ docs/CHANGELOG.md
@@ -33,7 +33,7 @@ All notable changes to this project will be documented in this file.
 
 | Status     | Count |
 | ---------- | ----- |
-| ✅ Passed  | 568   |
+| ✅ Passed  | 570   |
 | 🤔 Skipped | 0     |
 | 🔴 Failed  | 0     |
 | ⏱️ in      | 69s   |